+++ aDVanceERP.Modulos.CompraVenta/MVP/Modelos/DetalleCompraProducto.cs
@@ -5,7 +5,7 @@ namespace aDVanceERP.Modulos.CompraVenta.MVP.Modelos;
 public class DetalleCompraProducto : IEntidad {
     public DetalleCompraProducto() { }
 
-    public DetalleCompraProducto(long id, long idCompra, long idProducto, float cantidad, decimal precioCompra) {
+    public DetalleCompraProducto(long id, long idCompra, long idProducto, decimal cantidad, decimal precioCompra) {
         Id = id;
         IdCompra = idCompra;
         IdProducto = idProducto;
@@ -15,7 +15,7 @@ public class DetalleCompraProducto : IEntidad {
 
     public long IdCompra { get; set; }
     public long IdProducto { get; set; }
-    public float Cantidad { get; set; }
+    public decimal Cantidad { get; set; }
     public decimal PrecioCompra { get; set; }
 
     public long Id { get; set; }
