+++ aDVanceERP.Modulos.CompraVenta/MVP/Modelos/DetalleVentaProducto.cs
@@ -6,7 +6,7 @@ public class DetalleVentaProducto : IEntidad {
     public DetalleVentaProducto() { }
 
     public DetalleVentaProducto(long id, long idVenta, long idProducto, decimal precioCompraVigente,
-        decimal precioVentaFinal, float cantidad) {
+        decimal precioVentaFinal, decimal cantidad) {
         Id = id;
         IdVenta = idVenta;
         IdProducto = idProducto;
@@ -19,7 +19,7 @@ public class DetalleVentaProducto : IEntidad {
     public long IdProducto { get; set; }
     public decimal PrecioCompraVigente { get; set; }
     public decimal PrecioVentaFinal { get; set; }
-    public float Cantidad { get; set; }
+    public decimal Cantidad { get; set; }
 
     public long Id { get; set; }
 }
